chore: lib directory is moved to include one
