+++ endgame_evaluation.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include <iostream>
-#include "src/lib/board.hpp"
+#include "include/board.hpp"
 
 using namespace std;
