+++ include/cell_evaluation.cpp
@@ -1,5 +1,5 @@
 #include "cell_evaluation.hpp"
-#include "src/lib/board.hpp"
+#include "board.hpp"
 
 ArrStruct make_score() {
     ArrStruct a;
