+++ include/search.hpp
@@ -1,12 +1,12 @@
 #ifndef OTHELLO_TEXTBOOK_SEARCH_HPP
 #define OTHELLO_TEXTBOOK_SEARCH_HPP
 
-#include <iostream>
-#include <unordered_map>
-#include <vector>
-#include <algorithm>
-#include "src/lib/board.hpp"
-#include "src/lib/cell_evaluation.hpp"
+#include "iostream"
+#include "unordered_map"
+#include "vector"
+#include "algorithm"
+#include "board.hpp"
+#include "cell_evaluation.hpp"
 
 const int inf = 100000000;   // 大きな値
 const int cache_hit_bonus = 1000;   // 前回の探索で枝刈りされなかったノードへのボーナス
